CO-0015: Avoid memory leak with weak self and UIHelper
